move showDevelopmentOptions export out of react component
